requirements.txt
streamlit
yt-dlp
pillow
requests